+++ src/cli.rs
@@ -87,7 +87,8 @@ pub fn build_cli() -> Command<'static> {
                 .long("output")
                 .takes_value(true)
                 .value_hint(ValueHint::FilePath)
-                .help("Output file for non-colored ascii. If the output file is a plaintext file, no color will be used. The use color, either use a file with an .ansi extension, or an .html file, to convert the output to html."),
+                .help("Output file for non-colored ascii. If the output file is a plaintext file, no color will be used. The use color, either use a file with an .ansi extension, or an .html file, to convert the output to html.\
+                 .ansi files will consider environment variables when creating colored output, for example when COLORTERM is not set to truecolor, the resulting file will fallback to 8-bit colors."),
         )
         .arg(
             Arg::new("invert-density")