+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Improved handling of `.ansi`/`.ans` files ( includes more warnings)
+- Improved documentation of output files
+### Changed
+- Added `--release` flag to example build command (Thanks to @talwat)
+- Fixed failing doc test on rust nightly
 
 ## [1.0.3] - 2022-05-05